+++ lib/designs/whatsapp/whatsapp_sticker_editor.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 
 // Project imports:
 import 'package:pro_image_editor/modules/emoji_editor/emoji_editor.dart';
-import 'package:pro_image_editor/modules/sticker_editor.dart';
+import 'package:pro_image_editor/modules/sticker_editor/sticker_editor.dart';
 import 'package:pro_image_editor/utils/design_mode.dart';
 import '../../models/editor_configs/pro_image_editor_configs.dart';