+++ lib/modules/filter_editor/widgets/filtered_image.dart
@@ -7,14 +7,12 @@ import 'package:flutter/material.dart';
 // Project imports:
 import 'package:pro_image_editor/models/editor_image.dart';
 import 'package:pro_image_editor/pro_image_editor.dart';
-import '../../../models/history/filter_state_history.dart';
 import '../../../widgets/auto_image.dart';
-import '../utils/generate_filtered_image.dart';
+import '../types/filter_matrix.dart';
+import 'filter_generator.dart';
 
-/// Represents an image with multiple filters applied.
-///
-/// This widget displays an image with multiple filters applied on top of it. It also supports blur effect.
-class ImageWithFilters extends StatelessWidget {
+/// Represents an image where filters and blur factors can be applied.
+class FilteredImage extends StatelessWidget {
   /// The width of the image.
   final double width;
 
@@ -24,8 +22,8 @@ class ImageWithFilters extends StatelessWidget {
   /// The design mode of the image editor.
   final ImageEditorDesignModeE designMode;
 
-  /// The list of filter state histories to be applied on the image.
-  final List<FilterStateHistory> filters;
+  /// The list of filters to be applied on the image.
+  final FilterMatrix filters;
 
   /// The editor image to display.
   final EditorImage image;
@@ -36,7 +34,7 @@ class ImageWithFilters extends StatelessWidget {
   /// The blur factor
   final double blurFactor;
 
-  const ImageWithFilters(
+  const FilteredImage(
       {super.key,
       required this.width,
       required this.height,
@@ -56,19 +54,6 @@ class ImageWithFilters extends StatelessWidget {
       designMode: designMode,
     );
 
-    Widget filteredImg = img;
-    for (var filter in filters) {
-      filteredImg = generateFilteredImage(
-        child: filteredImg,
-        filter: filter.filter,
-        opacity: filter.opacity,
-      );
-    }
-
-    /*    print('-----------------------');
-    print(width);
-    print(height); */
-
     return SizedBox(
       width: width,
       height: height,
@@ -78,7 +63,7 @@ class ImageWithFilters extends StatelessWidget {
         alignment: Alignment.center,
         children: [
           img,
-          filteredImg,
+          ColorFilterGenerator(filters: filters, child: img),
           ClipRect(
             clipBehavior: Clip.hardEdge,
             child: BackdropFilter(
@@ -87,7 +72,7 @@ class ImageWithFilters extends StatelessWidget {
                 width: width,
                 height: height,
                 alignment: Alignment.center,
-                decoration: BoxDecoration(color: Colors.white.withOpacity(0.0)),
+                color: Colors.white.withOpacity(0.0),
               ),
             ),
           ),