+++ lib/modules/main_editor/utils/layer_interaction_manager.dart
@@ -13,7 +13,7 @@ import 'package:vibration/vibration.dart';
 import 'package:pro_image_editor/models/editor_configs/pro_image_editor_configs.dart';
 import 'package:pro_image_editor/models/theme/theme_layer_interaction.dart';
 import 'package:pro_image_editor/widgets/pro_image_editor_desktop_mode.dart';
-import '../../../models/history/last_position.dart';
+import '../../../models/history/last_layer_interaction_position.dart';
 import '../../../models/layer.dart';
 import '../../../utils/debounce.dart';