+++ lib/modules/paint_editor/paint_editor.dart
@@ -32,7 +32,7 @@ import '../../widgets/flat_icon_text_button.dart';
 import '../../widgets/platform_popup_menu.dart';
 import '../../widgets/pro_image_editor_desktop_mode.dart';
 import '../../widgets/transform/transformed_content_generator.dart';
-import '../filter_editor/widgets/image_with_filters.dart';
+import '../filter_editor/widgets/filtered_image.dart';
 import 'painting_canvas.dart';
 import 'utils/paint_desktop_interaction_manager.dart';
 import 'utils/paint_editor_enum.dart';
@@ -568,7 +568,7 @@ class PaintingEditorState extends State<PaintingEditor>
                               configs: configs,
                               transformConfigs:
                                   transformConfigs ?? TransformConfigs.empty(),
-                              child: ImageWithFilters(
+                              child: FilteredImage(
                                 width: getMinimumSize(
                                         mainImageSize, editorBodySize)
                                     .width,