+++ lib/pro_image_editor.dart
@@ -30,7 +30,7 @@ export 'utils/content_recorder.dart/utils/generate_high_quality_image.dart';
 
 export 'package:emoji_picker_flutter/emoji_picker_flutter.dart'
     show Emoji, RecentTabBehavior, CategoryIcons, Category, CategoryEmoji;
-export 'package:colorfilter_generator/presets.dart'
-    show presetFiltersList, PresetFilters;
-export 'package:colorfilter_generator/colorfilter_generator.dart';
-export 'package:colorfilter_generator/addons.dart';
+
+export 'modules/filter_editor/utils/filter_generator/filter_presets.dart';
+export 'modules/filter_editor/utils/filter_generator/filter_model.dart';
+export 'modules/filter_editor/utils/filter_generator/filter_addons.dart';