+++ test/modules/blur_editor_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 // Project imports:
 import 'package:pro_image_editor/models/init_configs/blur_editor_init_configs.dart';
-import 'package:pro_image_editor/modules/blur_editor.dart';
+import 'package:pro_image_editor/modules/blur_editor/blur_editor.dart';
 import '../fake/fake_image.dart';
 
 void main() {