+++ test/modules/sticker_editor_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 // Project imports:
-import 'package:pro_image_editor/modules/sticker_editor.dart';
+import 'package:pro_image_editor/modules/sticker_editor/sticker_editor.dart';
 import 'package:pro_image_editor/pro_image_editor.dart';
 
 void main() {